pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "article-pipeline"
version = "0.1.0"
description = "Article enrichment pipeline with OpenAI generation and local TranslateGemma translation."
requires-python = ">=3.13,<3.14"
dependencies = [
  "google-genai>=1.71.0",
  "huggingface-hub>=1.3.0",
  "json-repair>=0.58.7",
  "langgraph==0.6.7",
  "loguru==0.7.3",
  "mlx-lm>=0.31.0",
  "mlx-vlm>=0.4.3",
  "openai==1.109.1",
  "pydantic==2.12.5",
  "tqdm==4.67.1",
]

[project.scripts]
article-pipeline = "article_pipeline.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/article_pipeline", "src/local_translategemma", "src/local_gemma4_a4b"]
